+++ src/ui/buttons/index.js
@@ -1,4 +1,3 @@
 /* @flow */
 
 export * from './buttons';
-export * from './container';