+++ src/ui/index.js
@@ -1,3 +1,5 @@
 /* @flow */
 
 export * from './text';
+export * from './loading';
+export * from './buttons';