+++ src/zoid/checkout/component.jsx
@@ -6,17 +6,16 @@ import { node, dom } from 'jsx-pragmatic/src';
 import { getPayPalDomainRegex, getLogger, getLocale,
     getEnv, getClientID, getCommit, getSDKMeta, getCSPNonce, getBuyerCountry, getVersion } from '@paypal/sdk-client/src';
 import { ZalgoPromise } from 'zalgo-promise/src';
-import { create, CONTEXT, type ZoidComponent } from 'zoid/src';
+import { create, CONTEXT, type ZoidComponent, EVENT } from 'zoid/src';
 import { isDevice, memoize, noop, supportsPopups, inlineMemoize } from 'belter/src';
 import { FUNDING } from '@paypal/sdk-constants/src';
-import { Overlay } from '@paypal/common-components/src';
+import { Overlay, SpinnerPage } from '@paypal/common-components/src';
 
-import { getSessionID } from '../lib';
-import { DEFAULT_POPUP_SIZE, getCheckoutUrl } from '../config';
+import { getSessionID } from '../../lib';
+import { DEFAULT_POPUP_SIZE, getCheckoutUrl } from '../../config';
 
-import { containerContent } from './template/containerContent';
-import { componentTemplate } from './template';
 import type { CheckoutPropsType } from './props';
+import { containerContent } from './content';
 
 export function getCheckoutComponent() : ZoidComponent<CheckoutPropsType> {
     return inlineMemoize(getCheckoutComponent, () => {
@@ -38,7 +37,14 @@ export function getCheckoutComponent() : ZoidComponent<CheckoutPropsType> {
         
             logger: getLogger(),
         
-            prerenderTemplate: componentTemplate,
+            prerenderTemplate: ({ doc, props }) => {
+                return (
+                    <SpinnerPage
+                        nonce={ props.nonce }
+                    />
+                ).render(dom({ doc }));
+            },
+
             containerTemplate: ({ props, context, close, focus, doc, event, frame, prerenderFrame }) => {
                 const { locale: { lang } } = props;
                 const content = containerContent[lang];
@@ -165,7 +171,14 @@ export function getCheckoutComponent() : ZoidComponent<CheckoutPropsType> {
                     type:     'function',
                     required: false
                 },
-        
+
+                onFocused: {
+                    type:  'function',
+                    value: ({ event }) => {
+                        return (handler) => event.on(EVENT.FOCUS, handler);
+                    }
+                },
+                
                 test: {
                     type:    'object',
                     default: () => (window.__test__ || { action: 'checkout' })